requirements.txt
fastapi
uvicorn
python-docx
chromadb==0.4.24
groq
python-multipart
python-dotenv